refactor: add `infra/controller.js` layer and use in `/status` and `/migrations` to handle errors
